+++ main.py
@@ -1,38 +1,38 @@
-from game.scrabble import ScrabbleGame
+import os
+import sys
 
-def main():
-    print("Bienvenido a Scrabble Game")
-    try:
-        players_count = int(input("Ingrese la cantidad de jugadores: "))
+# Ruta absoluta del directorio raíz del proyecto
+project_root = os.path.dirname(os.path.abspath(__file__))
+# Agregar el directorio raíz a sys.path
+sys.path.insert(0, project_root)
+# Importar las clases de game sin problemas
+from game.scrabble import ScrabbleGame 
 
-        if players_count <= 1 or players_count > 4:
-            print("El número de jugadores debe estar entre 2 y 4.")
-        else:
-            # Realizar el juego para la cantidad de jugadores especificada
-            play_game(players_count)
-    except ValueError:
-        print("Por favor, ingrese un número válido de jugadores.")
-        return  # Terminar el programa si ocurre un error
-
-def play_game(players_count):
-    scrabble_game = ScrabbleGame(players_count)
-    print(f"Cantidad de Jugadores: {len(scrabble_game.players)}")
-
-    while scrabble_game.playing():
-        print(f"Turno del Jugador {scrabble_game.current_player.name}")
-        word = input("Ingrese Palabra: ")
-        location_x = int(input("Ingrese posición 'X': "))
-        location_y = int(input("Ingrese posición 'Y': "))
-        location = (location_x, location_y)
-        orientation = input("Ingrese la orientación (V/H): ").upper()
-
-        if scrabble_game.validate_word(word, location, orientation):
-            print("Palabra válida. ¡Bien hecho!")
-        else:
-            print("Palabra inválida. Inténtelo de nuevo.")
-
-        scrabble_game.next_turn()
-
+class Main:
+    def __init__(self):
+        print('¡Bienvenido a Scrabble Game!')
+        self.player_count = self.get_player_count()
+        self.game = ScrabbleGame(self.player_count)
+        
+    def valid_player_count(self, player_count):
+        try:
+            count = int(player_count)
+            return 2 <= count <= 4
+        except ValueError:
+            return False
+    
+    def get_player_count(self):
+        while True:
+            player_count = input('Cantidad de jugadores: ')
+            if self.valid_player_count(player_count):
+                return int(player_count)
+            print('Valor inválido')
+    
+    def play(self):
+        print(f'La cantidad de jugadores es: {self.player_count}')
+        self.game.next_turn()
+        print(f'Turno del jugador 1')
+        
 if __name__ == '__main__':
-    main()
-
+    main = Main()
+    main.play()